Rectify

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ int main(void)
 	long long int d;
 	float e;
 
-printf("size of a char: %zu byte(s)\n", sizeof(a));
-printf("size of int: %zu byte(s)\n", sizeof(b));
-printf("size of a long int: %zu byte(s)\n", sizeof(c));
-printf("size of a long long int: %zu byte(s)\n", sizeof(d));
+printf("size of a char: %lu byte(s)\n", (unsigned int) sizeof(a));
+printf("size of int: %lu byte(s)\n", (unsigned int) sizeof(b));
+printf("size of a long int: %lu byte(s)\n", (unsigned int) sizeof(c));
+printf("size of a long long int: %lu byte(s)\n", (unsigned int) sizeof(d));
 printf("size of a float: %zu byte(s)\n", sizeof(a));
 return ("0");
 }
